Consolidate all requirements file into one
There are currently no differences between mac and linux pip requirements,
and even if there were, they could be dealt with using platform markers
instead.

The linux one was split into two. I have tested this on ubuntu and it is
not a problem using a single one. This might have been a Cent OS thing,
so if problem comes up again, I suggest still keeping it as one file
but adding a for loop over the requirements file. I placed scipy first
for this reason (and floated numpy and cython up just for good measure).

I removed base requirements from the scenario requirements and changed
its name to reflect this.