SArray Construction from list perf fix
SArray Construction from list can create excessively large SFrame files
when the number of CPUs is very large compared to the size of the
SFrame. (For instance, 3000 rows, 64 threads).
This also affects SFrame construction from pandas.

This is because the SFrame is created with ncpu segments resulting in
highly inefficient compression. The fix is to simply construct the
SArray with 1 segment.