Allow reading of empty CSV files
It's a bit unintuitive to error out on `turicreate.SFrame.read_csv` on
an empty file with the message:

```
RuntimeError: No files corresponding to the specified path
```

Instead, let's return an empty SFrame. Removing some of the
empty-checking code seems to work and gives an empty SFrame with no
errors.