Generate compile_commands.json
This file is used with tooling like IDEs to get autocompletion and other
features that need to be aware of compile flags used on each source
file. Seems to affect build performance minimally, and only takes a few
MB of disk space when populated. I think it makes sense to turn it on
globally so that it's there when tools/IDEs need it.