Make explicit unsafe use of flexible_type
Adds 2 new methods to flexible_type

  reinterpret_get<T>
  reinterpret_mutable_get<T>

These two functions behave identically to get<T> and mutable_get<T> but
bypasses type checking. They are used in limited situations where it is
useful to reinterpret between integer and floating points types and to
make it clear when this is happening.

This also fixes the test failures introduced in #177