Fix TS_ASSERTS to be threadsafe
An annoying side effect is that TS_ASSERTS are now *really* slow.
In particular for some tests that loop TS_ASSERTS millions of times,
this will degrade test performance significantly. Those tests may have
to change from:

  TS_ASSERT(something);

to

  if (something) {
    TS_ASSERT(something);
  }

It will also be problematic if TS_ASSERT blocks, in particular, in
tests of threading primitives. For instance this is not a good idea.

  TS_ASSERT(wait_for_thread());

instead:

  int ret = wait_for_thread();
  TS_ASSERT(ret);