Fix C++ unit test failure, attempt #2 (#339)
The `readsome` function doesn't behave the way I expected - "available"
means "already in the buffer", not "can be read from the stream", which
means typically it reads 0 bytes on a newly initialized stream (which
breaks skip_BOM). We need to use `get`, and after each call, make sure
the stream is still in a good state (or recover if it's not).

This one passes both the C++ unity_sarray.cxxtest and Python
test_sframe.py.