Implemented exact sample and random_split
Implemented the following APIs:

    SFrame.random_split(..., exact=True)
    SFrame.sample(..., exact=True)
    SArray.sample(..., eaxct=True)

The inexact sample method simply does:

    threshold = 2^64 * fraction
    sample = sf[[0...size].hash(seed) < threshold]

The exact sample method does the same, but has to do a bit more work:

    random_numbers = [0...size].hash(seed)
    threshold = random_numbers.sort()[fraction * size]
    sample = sf[random_numbers < threshold]

Really, we only need a partial sort instead of a sort, but we do not
have a convenient implementation of partial sort to use. So this
implementation currently is quite expensive.

Fixes #347