Fix offset/length bug in NDArray-to-SArray
For np.ndarray input values which are slices of a base array, the conversion
procedure to flexible_type was using the entire buffer of the base array from
the offset onwards, instead of cropping to the length of the input slice. This
caused quadratic behavior, e.g., in the import of a value list(A), where A is
an np.ndarray of shape (N, 299, 299, 3).