Implementation of ndarray in ml_data.
This PR allows ndarray to be a
valid type that is understood by all the toolkits relying on ml_data.

Tests are added for exporting to CoreML, saving and loading, and other
implementations.  1d arrays following the ndarray path and vectors are
treated as interchangable.

Many other bugs were addressed along the way.