Merge commit of C API development effort
This commit exposes a C API to enable the direct wrapping of Turi Create
functionality by other languages.  Internal operations and structures, such as
SFrames, SArrays, and several ML models can now be used directly through an ABI
stable C interface.  This feature is designed to permit the use of Turi
Create as an embedded framework and facilitate the development of ML
applications in other languages besides C++ and python.

To build the C-API, run ./build_capi.sh in the base directory.  This
will generate a single dynamic library plus a single C header file that
allows complete interfacing to other languages.  Optionally, a Mac OSX
framework may be generated as well.

Currently, supported models are limited to the full supervised learning
suite; other models will be added in the future.

All SFrame and SArray operations are supported by the API.

In addition to the C-API, numerous under-the-hood improvements have been
made to allow for this interfacing, including:
- Updates to the extensions registration mechanism by which models are
exposed both to python and the C API.
- Updates to supervised learning so that it may be exposed through the
C API.
- Numerous small improvements and fixes to bugs exposed by testing the
code base through the C API.