Fixed critical bug in logistic/linear regression with validation tracking (#532)
* Fixed critical bug in logistic/linear regression when:
- Validation set is turned on
- Large number of features are present

This is escpecially a problem with image classification with large #
classes.

The second order information is being computed but never used. This is
an O(N * F * F * C * C) operation where F is the number of features
(which is around 1K) and N is the size of the validation set. This is
really, really expensive. This quadratic complexity is no longer needed.

* Fix to make the unit test pass.