Ternary Logic & and | operators (#589)
* Ternary Logic & and | operators

The current scheme of missing value propagation can lead to very
unexpected results. For instance,

    a = SArray([0,1,None])
    res = a[(a > 0) | (a == None)]

While I expect res to contain [1,None], this is currently not the case.
It will only return [1]. This is because "a > 0" propagates
None, and | propagates None.

This PR implements instead the three-value logic described in
https://en.wikipedia.org/wiki/Three-valued_logic#Kleene_and_Priest_logics

Essentially, instead of naive propagation of missing values, & and |
should following these truth tables, where '-' is "unknown".

and | F - T
-----------
  F | F F F
  - | F - -
  T | F - T

or | F - T
----------
 F | F - T
 - | - - T
 T | T T T

With these truth tables, the above code now works as expected.

This is a fundamental behavioral change and has a chance of breaking
people's code. On the other hand it is likely that if people's code
changes behavior with this PR, it is likely that their code is currently
not doing what they expect.

* Reverted change to CMakeLists for a more local change.