Fix build on 10.13
Restructures the use of `xcrun` in CMakeLists to make sure we are
setting appropriate environment variables for base SDK version/path,
then using the default path(s) to find libraries and frameworks. This
(combined with a fix from the iOS build for macOS as well) seems to make
framework header warnings get ignored as part of our build.

Fixes #638