[AC] Resolve #614 - train validation split bugs
In case of 'auto' validation set - the train validation split happened already
AFTER the dataset was pre-processed, hence the pre-processed data (which is
used for training) in fact included the validation data as well.

In addition - the split happened after the target was encoded, making it impossible
to filter the validation set by same target values as the training set (the user got
a crude, anclear error).