Comprehensively test JSON parsing using hypothesis
This adds a "test_arbitrary_json" test to SFrame JSON parsing.
Using the `hypothesis` module, we can generate test cases of
valid JSON, and compare the results of SFrame's parser to
Python's built-in JSON module.

This test uncovered 3 bugs, but they are documented in the
test as known issues and the test passes.