Remove "Turi Server" from documentation
There are several places we still refer to a "Turi Server" in the
API docs and user guide. This commit removes that wording (at this
point it's very confusing and not clear what it refers to), and in
some places, adds clarification that network filesystems should
work with SFrame/SArray.