Introduce platform-agnostic unity_neural_net library to wrap tcmps (#1144)
The primary change here is the introduction of unity_neural_net, which provides a minimal interface and builds on all platforms. Clients can safely depend on this library unconditionally, although for now it's only functional on Apple platforms, where it wraps the MPS-based backend. Eventually we will also wrap MXNet as well (adjusting the minimal interface as necessary).

Most of the changed files reflect a secondary change: renaming the turi::mps namespace to something more generic, turi::neural_net.