Implement initialization of pre-trained model for C-API object detector
The very first step of the `train` method will be to import the pre-trained (darknet) model parameters from a provided CoreML file. For now, this import is just a mapping from layer names to float_array values.

In addition, the object_detector class is structured in such a way as to facilitate dependency injection for proper unit testing.