Refactor vega_spec and data_spec munging into show method (#1155)
With this change, the C++ plot API matches the Python plot API,
which returns Vega-compatible spec and data JSON objects. Only
in the case of the show method (in either Python or C++) should
we prepend each with "vega_spec" or "data_spec" containers, as
those are specific to the way data is handled in tcviz. Prior
to this change, the outputs from C++ plot API were not
compatible with Vega without removing those containers.