Adopt CoreML data layout (NCHW) at library interfaces for network parameters (for object detection) (#1157)
We can push the MXNet-to-MPS-to-MXNet conversions into TCMPS, where this particular transpose can be implemented as a conversion between planar and "chunky" float images. In this way, we can standardize on the CoreML layouts where possible, for consistency, leaving our MPS usage of NHWC as an implementation detail.

Note that we're leaving the format for the actual network inputs as NHWC, since Object Detector data augmentation will need to be in HWC (interleaved RGB) anyway.