Fix to allow default arguments to extension types to work in python.
The type and quantity of the extenions arguments is checked rigorously
in the C++ part of the code after the default arguments have been filled
in, so the python check is not needed.  As the python check is not aware
of default arguments, it's better to just let the check happen in the
C++ layer.