Introduce neural_net::model_spec as primary storage for a learned NN
With this change, the C++ object_detector maintains (an encapsulated version of) a CoreML NeuralNetwork proto specifying the trained weights. This proto can be initialized with the pre-trained weights before training, and updated with the results of training when necessary. CoreML export will be able just to copy this proto into the output proto.