Allow empty plot labels in TC plots
Previously, empty string in TC plots was treated as a special
placeholder to indicate default value, and null values weren't
supported in plot APIs below Python.

With this change, null now means to omit the label, empty string
is actually empty string, and there is a __TC_DEFAULT_LABEL
placeholder string to indicate that the default should be used.