Fixes #1046. Placed resizing functionality from image_analysis to support conversion of 1-channel (Grayscale) and 4-channel (RGBA) images to 3-channel (RGB) images.
Currently supported color spaces (and their conversions to RGB) are Grayscale, palette, RGB, and RGBA.
PNG supports Grayscale, RGB and RGBA.
JPEG supports Grayscale and RGB, along with additional colorspaces such as CMYK, YCCK, and YCbCr. LibJpeg does not support conversion of YCCK or CMYK to RGB.
Since we expect image format to be a 3-channel RGB, I added an error check for the jpeg image to be in the RGB format. (Note: Conversion from CMYK to RGB can still happen using the resizing functionality but since it does so by dropping the K channel it seems wise to inform the user and have them feed the correct image without losing color information.)