Fix logger output in debug builds (#1235)
* Fix logger output in debug builds

Seems that perhaps we were relying on undefined behavior, and the
behavior has changed in newer compilers. (Unclear cause.)

The result is that in debug mode, the expression
`endltype(f) == endltype(std::endl)` always returns false,
even when the modifier passed in is in fact std::endl.
As a result, in debug mode, we never flush the stream.

This fix is to ignore which modifier is being passed in, and
always flush the stream as though it's std::endl. In practice,
it's likely to be endl most of the time, so the performance hit
from flushing more often should not be too bad.

* Add comment explaning the change