Fix plot get_data for SFrame summary view (#1252)
This change picks up two commits from a branch that were conceptually
necessary for other recent changes, but somehow didn't break the
build, or fail any tests, or break anything in practice except for
issue 1248. The necessary commits were:

https://github.com/znation/turicreate/commit/616de4b641389ce7ccc4e621850996af8d82781
https://github.com/znation/turicreate/commit/3e02db92ed685414f1be25f21bf23a0b9af3d32c

Fixes #1248

TBD: what specific commit went in that broke this? It worked in the 5.1
release.