Fix build in Xcode 9 (#1353)
For some reason, the version of clang++ in Xcode 9 gives warnings (as
error) like:

```
../src/unity/toolkits/object_detection/od_yolo.cpp:38:9: error: suggest
braces around initialization of subobject [-Werror,-Wmissing-braces]
      { 1, anchor_boxes.size(), (5 + num_classes), num_spatial });
        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
        {                                                     }
../src/unity/toolkits/object_detection/od_yolo.cpp:42:26: error: suggest
braces around initialization of subobject [-Werror,-Wmissing-braces]
                       { 0, 2, 1, 3 });
                         ^~~~~~~~~~
                         {         }
```

Not sure why later versions of Xcode don't give this warning, as it
seems to be valid and correct as far as I can tell (i.e. we should in
fact be using double braces in these places).