Link build status badge to branch overview
Currently, our build status badge links to whatever happens to be the "current" build, which is usually a random pull request build. This commit changes that to link to the branch overview page (which lists the default branch, master, separately and first). I couldn't find a better direct link for just the master branch.