Link build status badge to branch overview (#1467)
* Link build status badge to branch overview

Currently, our build status badge links to whatever happens to be the "current" build, which is usually a random pull request build. This commit changes that to link to the branch overview page (which lists the default branch, master, separately and first). I couldn't find a better direct link for just the master branch.

* Address review feedback

* Prevent build badge and Turi logo from being clickable (see https://stackoverflow.com/questions/40625614/is-it-possible-to-disable-the-automatic-linking-of-images-in-github-markdown-ren for workaround).
* Also adds two new badges for latest PyPI release version, and Python versions supported (this seems useful, at least on the level of build status).