Towards elimination of mutable_get.
Eliminating mutable_get access to flexible_type will enable us to better
use memory pools and reference counting to reduce allocations and
copying.  In this PR, it's changed to a private method, and instead a
function, get_and_reset<...>(), returns the object using a move
operation if it's unique and a copy if it's not.