Fixes CSV parsing for unmatched quotes extending past last line (#1694)
Addresses #1692.

The newline delimiting logic when quote characters are used works
by computing the "quote parity" for every new line character, accepting
only new line characters where quote parity is False. This permits
parsing of files of the form

```
 user_id, review
 123, "good food"
 456, "hello
 world"
 678, "moof"
```

To simplify end of file handling, a new line character is always
inserted into the buffer if not already there when EOF is hit. However,
if there is a mismatched quotechar, this is not recognized as a valid
new line resulting in lines being left in the buffer and ignored
entirely.

The patch is simple, the final new line character at EOF is *always*
a valid new line (flagged with quote_parity=False)