Fix broken build with `--no-python`
The big CMake refactor that went in recently seems to have broken
the build for non-Python builds (./configure --no-python). The problem
is that unity_shared now only builds in the python directory (which
makes sense, since it's only used in Python), but the extensions all
depended on it and all were expected to build even without Python.

The fix is to move the `unity_shared` target back up from
`src/unity/python/turicreate` into `src/unity`. Now it builds with
non-Python builds again.