Remove storyboard from TCViz app (#1873)
We've had some issues with the storyboard file since the existence of
this project:

* Difficult to merge changes
* Unclear/conflicting sources of truth (code vs. storyboard)
* Difficulty compiling on various platforms/versions of Xcode

Given that we have a very simple UI (a light wrapper around WKWebView),
there isn't really a need for a storyboard. The equivalent can be easily
created in code, and will be easier to modify and maintain in the long run.