Speed up image row rendering in evaluation.explore (#2015)
Previously, we would show _all_ images in the browser (whether or not the
user could actually see them, since they would be hidden behind a
scrollable area). This isn't actually necessary, since the user can't
really act on this much information. With this commit, images are now
limited to:

* 20 images per row in the table that shows examples from each class in
  the test set.
* 100 images per confusion (actual/expected label pair) in the
  drill-down view.

Fixes #1940