Fixes a hang in CSV parsing when parsing a token of a single '[' (#2062)
The lost lasting issue is that there doesn't seem to be an easy way to
disble the skip parser in boost spirit. A hack is to use qi::eoi as the
skip parser, but that is not ideal; under certain circumstances when
the parser calls the skip parser at the end of the line, this can cause a hang
since the eoi parser does not consume any characters.

What seems to work is to use !eps as the parser. eps just returns true.
!eps thus returns false. And that works to fully disable the skip
parser.