Fix for broken unit test in vega renderer (#2188)
Allows arbitrary tag names without asserting false -- if we need to
catch an unintended tag name here, we can use LogProxy (added in #2180)
to debug it. Otherwise, the vega example tests try to create span
elements and fail due to the assertion. The tests pass as is without the
assertion.

Fixes #2186