Cleaned up the use of --no-remotefs flags. (#2387)
Cleaned up the use of --no-remotefs flags.

Previously, due to some poor logic, this flag still allowed compilation of a number of remotefs specific packages, and several operations on downloading things were similarly problematic.  This change corrects this logic.  

Now, in CMakeLists.txt files, the ${TC_BUILD_REMOTEFS} returns 1 if remote fs is enabled, which is the default.  In code, the macro TC_DISABLE_REMOTEFS is defined if and only if there is no remotefs support.

An additional docker test was also added to make sure things compile and all cpp tests run with this support.