Build System Improvements (#2835)
This Commit introduces a significant overhaul of the build system with a focus on simplification and integration with proper CMake options.

Improvements:

All dependencies except openssl, curl, and libxml are moved into src/external and compiled as part of our source build. These dependencies are those needed for adding remote fs support, so compiling with --no-remotefs does not compile anything in deps/.

CMakeLists.txt simplification. Now CMakeLists.txt is organized in a far more standard way. In addition, more logic is added to and moved to the ./configure script, as many cmake options that we were previously overriding with our own compiler flags work correctly when specified in CMake BEFORE the CMakeLists.txt file is loaded.

More options in ./configure to support additional compilation options, such as listing out all source files used in the build or turning on or off various options.

The custom model dylibs for iOS are now compiled using xcodebuild. XCode is supported as a generator for these models, which means that the compilation builds an xcode project, then calls xcodebuild to compile it. This allows better integration with the OSX ecosystem and should provide a way to resolve #834 easily.

Source changes needed to support the above compilation and other internal dependencies. These changes included proper use of ARC globally, fixing all the size_t -> int conversions (int is 32 bit on some platforms, while size_t is 64 bit).