Pipelining API + Pipelining for Image Feature Extraction (#2918)
This PR adds a run_background API to run extension functions in the background.  It's used to improve support for pipelining in image feature extraction.

Many of the current speed regressions could be addressed with proper pipelining support in python. This PR adds the ability to run extension functions on a background thread while the main python thread does other things. This is done by using a futures interface that works on top of extension functions and can be called using a run_background method on the extension function call.

Previously, we could run the following code:
```
for i in range(n):
    prepared_data = tc.extensions.prepared_data(data_info, i)
    run_tf_model(prepared_data)
```

The problem is that this is run in serial, meaning the CPU and GPU are not both busy. With this API, you can replace this with:

```
# Init for the first iteration here
data_future = tc.extensions.prepared_data.run_background(data_info, 0)

for i in range(n):
    # block until background function finishes; get result.
    prepared_data = data_future.wait()

    # start data prep for next index in background
    if i + 1 != n:
        data_future = tc.extensions.prepared_data.run_background(data_info, i + 1)
   
    run_tf_model(prepared_data)
```

This should make pipelining support much easier.

In addition, implementing this with image feature extraction gets the speed to within  10% of the 5.8 speed on both CPU and GPU. Resolves https://github.com/apple/turicreate/issues/2813.