Introduce a minimal reactive streams framework, inspired somewhat by Swift's Combine
The main goal is make it easier to reason about the various portions of our
deep-learning training pipelines. For example, in Object Detection, we have a
data iterator that produces raw training examples, an image augmenter that
performs data augmentation, and a model backend that consumes a stream of
augmented training examples and produces progress updates and models.

As we integrate this framework, it should be easier to maintain unit tests for
these evolving pipelines and to manipulate threading and concurrency. For
example, if done right, this can replace our implementations of "double
buffering" currently copied across each model toolkit.